+++ plugins/org.yakindu.base.expressions/emf-gen/org/yakindu/base/expressions/expressions/impl/RealLiteralImpl.java
@@ -33,7 +33,7 @@ public class RealLiteralImpl extends LiteralImpl implements RealLiteral {
 	 * @generated
 	 * @ordered
 	 */
-	protected static final float VALUE_EDEFAULT = 0.0F;
+	protected static final double VALUE_EDEFAULT = 0.0;
 
 	/**
 	 * The cached value of the '{@link #getValue() <em>Value</em>}' attribute.
@@ -43,7 +43,7 @@ public class RealLiteralImpl extends LiteralImpl implements RealLiteral {
 	 * @generated
 	 * @ordered
 	 */
-	protected float value = VALUE_EDEFAULT;
+	protected double value = VALUE_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -69,7 +69,7 @@ public class RealLiteralImpl extends LiteralImpl implements RealLiteral {
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public float getValue() {
+	public double getValue() {
 		return value;
 	}
 
@@ -78,8 +78,8 @@ public class RealLiteralImpl extends LiteralImpl implements RealLiteral {
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setValue(float newValue) {
-		float oldValue = value;
+	public void setValue(double newValue) {
+		double oldValue = value;
 		value = newValue;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, ExpressionsPackage.REAL_LITERAL__VALUE, oldValue, value));
@@ -108,7 +108,7 @@ public class RealLiteralImpl extends LiteralImpl implements RealLiteral {
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case ExpressionsPackage.REAL_LITERAL__VALUE:
-				setValue((Float)newValue);
+				setValue((Double)newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
